Extract the routers to add the proper one for each rails version
Rails 3.0 and 3.1 use Rack::Mount, Rails 3.2 uses the new Journey
Router
